main: split binary protocol into proto_bin.c
also some missed text bits.
