+++ Makefile.am
@@ -26,7 +26,8 @@ memcached_SOURCES = memcached.c memcached.h \
                     slab_automove.c slab_automove.h \
                     authfile.c authfile.h \
                     restart.c restart.h \
-                    proto_text.c proto_text.h
+                    proto_text.c proto_text.h \
+                    proto_bin.c proto_bin.h
 
 if BUILD_SOLARIS_PRIVS
 memcached_SOURCES += solaris_priv.c
