+++ proto_text.h
@@ -2,7 +2,7 @@
 #define PROTO_TEXT_H
 
 /* text protocol handlers */
-void process_command(conn *c, char *command);
+void complete_nread_ascii(conn *c);
 int try_read_command_asciiauth(conn *c);
 int try_read_command_ascii(conn *c);
